添加model，操作数据库，未写完
