+++ .gitignore
@@ -13,12 +13,12 @@
 
 # Dependency directories (remove the comment below to include it)
 # vendor/
-app/user/config/config.yml
-app/task/config/config.yml
+user_service/app/user/config/config.yml
+user_service/app/task/config/config.yml
 config/config.yml
-app/gateway/logs
-app/user/logs
-app/task/logs
+user_service/app/gateway/logs
+user_service/app/user/logs
+user_service/app/task/logs
 .idea
 bin
 logs
